+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/typeresolution/rules/imports/UnusedImports.java
@@ -3,10 +3,11 @@
  */
 package net.sourceforge.pmd.lang.java.typeresolution.rules.imports;
 
+import java.util.Iterator;
+
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.java.ast.ASTImportDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTName;
-import net.sourceforge.pmd.lang.java.ast.DummyJavaNode;
 import net.sourceforge.pmd.lang.java.ast.TypeNode;
 import net.sourceforge.pmd.lang.java.rule.imports.UnusedImportsRule;
 import net.sourceforge.pmd.lang.rule.ImportWrapper;
@@ -17,7 +18,7 @@ public class UnusedImports extends UnusedImportsRule {
     public Object visit(ASTImportDeclaration node, Object data) {
 	if (node.isImportOnDemand()) {
 	    ASTName importedType = (ASTName) node.jjtGetChild(0);
-	    imports.add(new ImportWrapper(importedType.getImage(), null, node));
+	    imports.add(new ImportWrapper(importedType.getImage(), null, node, node.getType(), node.isStatic()));
 	} else {
 	    super.visit(node, data);
 	}
@@ -30,14 +31,18 @@ public class UnusedImports extends UnusedImportsRule {
 	    return;
 	}
 	ImportWrapper candidate = getImportWrapper(node);
-	if (imports.contains(candidate)) {
-	    imports.remove(candidate);
-	    return;
+	Iterator<ImportWrapper> it = imports.iterator();
+	while (it.hasNext()) {
+	    ImportWrapper i = it.next();
+	    if (i.matches(candidate)) {
+	        it.remove();
+	        return;
+	    }
 	}
 	if (TypeNode.class.isAssignableFrom(node.getClass()) && ((TypeNode) node).getType() != null) {
 	    Class<?> c = ((TypeNode) node).getType();
 	    if (c.getPackage() != null) {
-		candidate = new ImportWrapper(c.getPackage().getName(), null, new DummyJavaNode(-1));
+		candidate = new ImportWrapper(c.getPackage().getName(), null);
 		if (imports.contains(candidate)) {
 		    imports.remove(candidate);
 		}